Reorganized lib folder
