+++ src/lib/Miscellaneous/EditIcon.svelte
